split out ui code to simplify main class
